Add abstract user string condition type and refactor email/username conditions
